requirements.txt
contourpy==1.2.0
cycler==0.12.1
dill==0.3.8
fonttools==4.50.0
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pdoc==14.4.0
pillow==10.2.0
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
setuptools==68.2.2
six==1.16.0
tqdm==4.66.2
typing-filter==0.1.1
wheel==0.41.2
